Add files via upload
BA Project tries to use a web-linked .json file as the data source.
The Copy uses a downloaded version of the .csv file.
The .json file is very incomplete (only capturing about 1000 rows of 600,000 entries)
The last 3 columns in the json have their names corrupted, and that's fixed.
The .csv has all rows with a Null removed. This may change in the future.